Separate category update test to __test__/category #13
